Remove redundant __subclasshook__ methods, use ABC insted of ABCMeta
The subclass hooks are not required because ABC will prevent instantiation
of a class unless all of the abstract methods are implemented.

The change from using ABCMeta meta class to ABC is stylistic in nature
only, but it's arguably clearer to a reader who is not familiar with
meta classes.